+++ agentic_mail_mcp/Gmail/Domain/ValueObjects/__init__.py
@@ -1,3 +1,4 @@
+from agentic_mail_mcp.Gmail.Domain.ValueObjects.attached_message import AttachedMessage
 from agentic_mail_mcp.Gmail.Domain.ValueObjects.email_address import EmailAddress
 from agentic_mail_mcp.Gmail.Domain.ValueObjects.gmail_message_id import GmailMessageId
 from agentic_mail_mcp.Gmail.Domain.ValueObjects.gmail_query import GmailQuery
@@ -5,6 +6,7 @@ from agentic_mail_mcp.Gmail.Domain.ValueObjects.history_id import HistoryId
 from agentic_mail_mcp.Gmail.Domain.ValueObjects.thread_id import ThreadId
 
 __all__ = [
+    "AttachedMessage",
     "EmailAddress",
     "GmailMessageId",
     "GmailQuery",